small OpenGL implementation, abstracted header file
