+++ .gitignore
@@ -30,3 +30,7 @@
 *.exe
 *.out
 *.app
+
+# Other
+bin/
+*.geany
